Various improvements
# RadioStream
- RadioStream now uses proper convention to denote private methods and variables
- RadioStream can now accept an int `preroll_len` at init
- RadioStream will now throw an exception if it runs out of attempts to get a stream URL
- RadioStream now dumps the first self._preroll_len bytes when opening a stream
- Added read-only properties for byte_buffer and url in RadioStream

# Added RadioStreamManager
This class creates multiple RadioStreams and monitors them. It will denote one stream as the primary stream. If a non-primary stream fails, it is removed and a new one created to fill its space. If the primary stream fails, it is replaced by a functioning non-primary stream. If all streams fail, a new stream is created and used as the primary stream, then the non-primary streams are recreated.

When the primary stream is replaced, RadioStreamManager invokes all its on_stream_failover_handler callables with the old primary and the new primary streams as arguments.

# Added RedundantRadioStream
Uses a RadioStreamManager to start multiple streams and write the primary's contents to its own buffer. If the primary fails, it syncs the new primary as best it can.