Now, after hitting the play button, wait for either the stream to start OR a streaming URL to appear.
The streaming URL doesn't appear in the performance metrics until the connection is closed.
That means, if the stream is actually trying to start, we need to wait for it to start, then hit the stop button.
But if the stream is never going to start (e.g., no compaitible audio codecs or no audio sinks present), then we don't want to wait 30 seconds for no reason.
In that second case, the stream URL appears in the performance metrics, so we can just check for that at the same time as the normal stream start metrics to avoid waiting for no reason.